+++ lib/ReactViews/DragDropFile.tsx
@@ -1,8 +1,8 @@
-import classNames from "classnames";
-import { action, runInAction, makeObservable } from "mobx";
+import { action, runInAction } from "mobx";
 import { observer } from "mobx-react";
-import { DragEvent, MouseEvent, Component } from "react";
-import { Trans, withTranslation, WithTranslation } from "react-i18next";
+import { DragEvent, FC, MouseEvent, useRef, useState } from "react";
+import { Trans } from "react-i18next";
+import styled, { useTheme } from "styled-components";
 import {
   Category,
   DataSourceAction
@@ -13,27 +13,57 @@ import CatalogMemberMixin, { getName } from "../ModelMixins/CatalogMemberMixin";
 import MappableMixin from "../ModelMixins/MappableMixin";
 import addUserFiles from "../Models/Catalog/addUserFiles";
 import { BaseModel } from "../Models/Definition/Model";
-import Styles from "./drag-drop-file.scss";
-import { WithViewState, withViewState } from "./Context";
+import Box from "../Styled/Box";
+import Icon, { StyledIcon } from "../Styled/Icon";
+import Text from "../Styled/Text";
 import { raiseFileDragDropEvent } from "../ViewModels/FileDragDropListener";
+import { useViewState } from "./Context";
+import DragDropNotification from "./DragDropNotification";
 
-interface PropsType extends WithTranslation, WithViewState {}
+const DropZone = styled.div<{ isActive: boolean }>`
+  position: absolute;
+  background: transparent;
+  opacity: 0;
+  z-index: -1;
+  transition: height 0.25s ease-in-out;
 
-@observer
-class DragDropFile extends Component<PropsType> {
-  target: EventTarget | undefined;
+  ${({ isActive, theme }) =>
+    isActive &&
+    `
+    position: fixed;
+    top: 0;
+    left: 0;
 
-  constructor(props: PropsType) {
-    super(props);
-    makeObservable(this);
-  }
+    width: 100vw;
+    height: 100vh;
+    z-index: 99999;
 
-  async handleDrop(e: DragEvent) {
+    transition: opacity, 0.2s;
+
+    opacity: 1;
+    display: block;
+    border: 0;
+
+    background: radial-gradient(
+      ellipse at center,
+      ${theme.transparentDark} 0%,
+      ${theme.dark} 100%
+    );
+  `}
+`;
+
+const DragDropFile: FC = observer(() => {
+  const viewState = useViewState();
+  const theme = useTheme();
+  const targetRef = useRef<EventTarget>();
+  const [lastUploadedFiles, setLastUploadedFiles] = useState<readonly string[]>(
+    []
+  );
+
+  const handleDrop = async (e: DragEvent) => {
     e.persist();
     e.preventDefault();
     e.stopPropagation();
-
-    const props = this.props;
 
     for (let i = 0; i < e.dataTransfer.files.length; i++) {
       // Log event to analytics for each file dropped (sometimes multiple files dropped in one DragEvent)
@@ -41,7 +71,7 @@ class DragDropFile extends Component<PropsType> {
         e.dataTransfer.files[i].type ||
         e.dataTransfer.files[i].name.split(".").pop(); // use file extension if type property is empty
 
-      this.props.viewState.terria.analytics?.logEvent(
+      viewState.terria.analytics?.logEvent(
         Category.dataSource,
         DataSourceAction.addFromDragAndDrop,
         `File Type: ${fileType}, File Size(B): ${e.dataTransfer.files[i].size}`
@@ -51,27 +81,26 @@ class DragDropFile extends Component<PropsType> {
     try {
       const addedCatalogItems: BaseModel[] | undefined = await addUserFiles(
         e.dataTransfer.files,
-        props.viewState.terria,
-        props.viewState
+        viewState.terria,
+        viewState
       );
 
       if (isDefined(addedCatalogItems) && addedCatalogItems.length > 0) {
-        runInAction(() => (props.viewState.myDataIsUploadView = false));
-        if (props.viewState.explorerPanelIsVisible) {
+        runInAction(() => (viewState.myDataIsUploadView = false));
+        if (viewState.explorerPanelIsVisible) {
           (
-            await props.viewState.viewCatalogMember(addedCatalogItems[0])
+            await viewState.viewCatalogMember(addedCatalogItems[0])
           ).throwIfError(`Failed to view ${getName(addedCatalogItems[0])}`);
-          props.viewState.openUserData();
+          viewState.openUserData();
         } else {
           // update last batch of uploaded files
-          runInAction(
-            () =>
-              (props.viewState.lastUploadedFiles = addedCatalogItems.map(
-                (item) =>
-                  CatalogMemberMixin.isMixedInto(item)
-                    ? item.name
-                    : item.uniqueId
-              ))
+          setLastUploadedFiles(
+            addedCatalogItems.map(
+              (item) =>
+                (CatalogMemberMixin.isMixedInto(item)
+                  ? item.name
+                  : item.uniqueId) as string
+            )
           );
         }
 
@@ -83,7 +112,7 @@ class DragDropFile extends Component<PropsType> {
         Result.combine(
           await Promise.all(mappableItems.map((f) => f.loadMapItems())),
           "Failed to load uploaded files"
-        ).raiseError(props.viewState.terria);
+        ).raiseError(viewState.terria);
 
         raiseFileDragDropEvent({
           addedItems: mappableItems,
@@ -97,72 +126,79 @@ class DragDropFile extends Component<PropsType> {
 
         if (isDefined(firstZoomableItem)) {
           runInAction(() =>
-            props.viewState.terria.currentViewer.zoomTo(firstZoomableItem, 1)
+            viewState.terria.currentViewer.zoomTo(firstZoomableItem, 1)
           );
         }
       }
 
-      runInAction(() => (props.viewState.isDraggingDroppingFile = false));
+      runInAction(() => (viewState.isDraggingDroppingFile = false));
     } catch (e) {
-      props.viewState.terria.raiseErrorToUser(e, "Failed to upload files");
+      viewState.terria.raiseErrorToUser(e, "Failed to upload files");
     }
-  }
+  };
 
-  @action
-  handleDragEnter(e: DragEvent) {
+  const handleDragEnter = action((e: DragEvent) => {
     e.preventDefault();
     e.stopPropagation();
     e.dataTransfer.dropEffect = "copy";
-    this.target = e.target;
-  }
+    targetRef.current = e.target;
+  });
 
-  handleDragOver(e: DragEvent) {
+  const handleDragOver = (e: DragEvent) => {
     e.preventDefault();
-  }
+  };
 
-  @action
-  handleDragLeave(e: MouseEvent) {
+  const handleDragLeave = action((e: MouseEvent) => {
     e.preventDefault();
     if (e.screenX === 0 && e.screenY === 0) {
-      this.props.viewState.isDraggingDroppingFile = false;
+      viewState.isDraggingDroppingFile = false;
     }
-    if (e.target === document || e.target === this.target) {
-      this.props.viewState.isDraggingDroppingFile = false;
+    if (e.target === document || e.target === targetRef.current) {
+      viewState.isDraggingDroppingFile = false;
     }
-  }
+  });
 
-  @action
-  handleMouseLeave() {
-    this.props.viewState.isDraggingDroppingFile = false;
-  }
+  const handleMouseLeave = action(() => {
+    viewState.isDraggingDroppingFile = false;
+  });
 
-  render() {
-    return (
-      <div
-        onDrop={this.handleDrop.bind(this)}
-        onDragEnter={this.handleDragEnter.bind(this)}
-        onDragOver={this.handleDragOver.bind(this)}
-        onDragLeave={this.handleDragLeave.bind(this)}
-        onMouseLeave={this.handleMouseLeave.bind(this)}
-        className={classNames(Styles.dropZone, {
-          [Styles.isActive]: this.props.viewState.isDraggingDroppingFile
-        })}
+  return (
+    <>
+      <DropZone
+        isActive={viewState.isDraggingDroppingFile}
+        onDrop={handleDrop}
+        onDragEnter={handleDragEnter}
+        onDragOver={handleDragOver}
+        onDragLeave={handleDragLeave}
+        onMouseLeave={handleMouseLeave}
       >
-        {this.props.viewState.isDraggingDroppingFile ? (
-          <div className={Styles.inner}>
+        {viewState.isDraggingDroppingFile ? (
+          <Box centered fullWidth fullHeight column>
+            <StyledIcon glyph={Icon.GLYPHS.dragDrop} styledHeight="80px" />
             <Trans i18nKey="dragDrop.text">
-              <h3 className={Styles.heading}>Drag & Drop</h3>
-              <div className={Styles.caption}>
+              <Text
+                textLight
+                textAlignCenter
+                pop
+                as="h3"
+                css={`
+                  margin-top: ${theme.spacing}px;
+                  font-weight: 700;
+                `}
+              >
+                Drag & Drop
+              </Text>
+              <Text textLight pop textAlignCenter styledFontSize="24px">
                 Your data anywhere to view on the map
-              </div>
+              </Text>
             </Trans>
-          </div>
-        ) : (
-          ""
-        )}
-      </div>
-    );
-  }
-}
+          </Box>
+        ) : null}
+      </DropZone>
+      <DragDropNotification uploadedFiles={lastUploadedFiles} />
+    </>
+  );
+});
+DragDropFile.displayName = "DragDropFile";
 
-export default withTranslation()(withViewState(DragDropFile));
+export default DragDropFile;